# docs/05_APIs/zentraos-openapi.json
{
  "components": {
    "schemas": {
...
      "ChatMessageRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "message": "Why did EU refunds increase from June to July 2026?"
          }
        ],
        "properties": {
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatMessageRequest",
        "type": "object"
      },
      "ChatPageResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ThreadSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ChatPageResponse",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "examples": [
          {
            "actions": {
              "can_append_message": true,
              "can_archive": true,
              "can_cancel": false,
              "can_delete": true,
              "can_restore": false,
              "can_retry": false
            },
            "created_at": "2026-08-01T09:00:00Z",
            "latest_activity_at": "2026-08-01T09:00:00Z",
            "messages": [
              {
                "authored_by_user": true,
                "content": "How is the business doing?",
                "created_at": "2026-08-01T09:00:00Z",
                "kind": "user_question",
                "message_id": "44000000-0000-0000-0000-000000000001"
              },
              {
                "authored_by_user": false,
                "content": "I could not map that message to a governed question. Please choose or rephrase one supported question.\n- Why did EU refunds increase from June to July 2026?\n- Which sales channel accounted for the increase in North America revenue from October to November 2026?",
                "created_at": "2026-08-01T09:00:00Z",
                "kind": "router_clarification",
                "message_id": "44000000-0000-0000-0000-000000000002"
              }
            ],
            "project_id": "42000000-0000-0000-0000-000000000001",
            "routing": {
              "clarification": "I could not map that message to a governed question. Please choose or rephrase one supported question.",
              "disposition": "unsupported",
              "suggestions": [
                "Why did EU refunds increase from June to July 2026?",
                "Which sales channel accounted for the increase in North America revenue from October to November 2026?"
              ]
            },
            "status": "draft",
            "thread_id": "43000000-0000-0000-0000-000000000001",
            "title": "How is the business doing?",
            "updated_at": "2026-08-01T09:00:00Z",
            "usage": {
              "input_tokens": 0,
              "latency_ms": 0,
              "output_tokens": 0,
              "total_cost_usd": "0"
            }
          }
        ],
        "properties": {
          "actions": {
            "$ref": "#/components/schemas/ThreadActionsResponse"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_cursor": {
            "default": 0,
            "title": "Event Cursor",
            "type": "integer"
          },
          "investigation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigation Id"
          },
          "investigations": {
            "items": {
              "$ref": "#/components/schemas/ThreadInvestigationResponse"
            },
            "title": "Investigations",
            "type": "array"
          },
          "latest_activity_at": {
            "format": "date-time",
            "title": "Latest Activity At",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ThreadMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageResponse"
          }
        },
        "required": [
          "thread_id",
          "project_id",
          "title",
          "status",
          "created_at",
          "updated_at",
          "latest_activity_at",
          "messages",
          "usage",
          "actions"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ThreadActionsResponse": {
        "additionalProperties": false,
        "properties": {
          "can_append_message": {
            "title": "Can Append Message",
            "type": "boolean"
          },
          "can_archive": {
            "title": "Can Archive",
            "type": "boolean"
          },
          "can_cancel": {
            "title": "Can Cancel",
            "type": "boolean"
          },
          "can_delete": {
            "title": "Can Delete",
            "type": "boolean"
          },
          "can_restore": {
            "title": "Can Restore",
            "type": "boolean"
          },
          "can_retry": {
            "title": "Can Retry",
            "type": "boolean"
          }
        },
        "required": [
          "can_append_message",
          "can_archive",
          "can_restore",
          "can_delete",
          "can_cancel",
          "can_retry"
        ],
        "title": "ThreadActionsResponse",
        "type": "object"
      },
...
      "ThreadMessageResponse": {
        "additionalProperties": false,
        "properties": {
          "authored_by_user": {
            "title": "Authored By User",
            "type": "boolean"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "kind",
          "content",
          "created_at",
          "authored_by_user"
        ],
        "title": "ThreadMessageResponse",
        "type": "object"
      },
...
      "UsageResponse": {
        "additionalProperties": false,
        "properties": {
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "total_cost_usd": {
            "title": "Total Cost Usd",
            "type": "string"
          }
        },
        "required": [
          "input_tokens",
          "output_tokens",
          "total_cost_usd",
          "latency_ms"
        ],
        "title": "UsageResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Nexus API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chats/{chat_id}": {
      "delete": {
        "operationId": "delete_chat_v1_chats__chat_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Chat"
      },
      "get": {
        "operationId": "get_chat_v1_chats__chat_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chat"
      }
    },
    "/v1/chats/{chat_id}/archive": {
      "post": {
        "operationId": "archive_chat_v1_chats__chat_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Chat"
      }
    },
    "/v1/chats/{chat_id}/events": {
      "get": {
        "operationId": "stream_chat_events_v1_chats__chat_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Chat Events"
      }
    },
    "/v1/chats/{chat_id}/messages": {
      "post": {
        "operationId": "append_chat_message_v1_chats__chat_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Chat Message"
      }
    },
    "/v1/chats/{chat_id}/restore": {
      "post": {
        "operationId": "restore_chat_v1_chats__chat_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Chat"
      }
    },
...
    "/v1/groups/{group_id}/chats": {
      "get": {
        "operationId": "list_chats_v1_groups__group_id__chats_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chats"
      },
      "post": {
        "operationId": "create_chat_v1_groups__group_id__chats_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Chat"
      }
    },
...
  }
}